refactor(fasting,widgets): a tab out of a 1,779-line screen, and the last two unseamed calculators
Audit §5.1. `FastTrackerScreen.kt` was 1,779 lines with 15 private composables. The registry
records exactly this shape being fixed for Khatam — "14 inline private composables across the
Khatam screens collapsed into 4 shared components" — and this one file was larger than all
the Khatam screens combined.

## The tab that was a screen

`MakeupFastsContent` plus its four helpers is ~340 lines: what is owed, what has been paid as
fidya, and what is done. `Route.MakeupFasts` was deleted because make-up fasts is a tab
inside `FastTrackerScreen`, and that is still true — but a tab this size belongs in a file
where it can be read and previewed on its own terms. `MakeupFastsTab.kt`, cut and pasted, with
visibility widened from `private` to `internal` and nothing else changed. 1,779 -> 1,434.

## The calculation that was in a screen

`calculateAyyamAlBeedDays` — how many days until the white days, the 13th to 15th of the Hijri
month — was a private function in that screen file, which gave it three problems:

  1. No test could reach it. It has six now, including one that walks a year asserting the
     answer never goes negative or exceeds a lunar month.
  2. One of its two call sites passed `LocalDate.now()`, read at composition. A screen open
     across midnight counted down to yesterday's answer for as long as it stayed open. It
     reads `TodayProvider` now — the seam the registry introduced for exactly this, and the
     reason the test can say what day it is.
  3. It ignored the user's `hijriDayOffset` while event matching honoured it, so someone who
     had shifted their Hijri date by a day to match a local moon sighting saw the two
     disagree. That is registry Open #10, and this is one more helper it stops being true of.

It is `GetDaysUntilAyyamAlBeedUseCase` in `domain/usecase/fasting/`, and `FastingViewModel`
publishes the result in `RamadanTrackerUiState`.

The offset reaches it through a new `HijriSettings` seam — one `Flow<Int>` — rather than the
whole 179-member `SettingsRepository`, which is the argument `ZakatSettings` already settled
for the zakat currency (#436). Three unrelated readers want that one flow.

## The last direct `PrayerTimeCalculator()`

`WidgetsScreen`'s preview builder constructed one, the last site of a pattern the checklist
had already removed from five ViewModels. It is injected now, through `SettingsViewModel`.
`grep -rn "PrayerTimeCalculator()" app/src/main` returns nothing.

Docs: `ARCHITECTURE.md` §9 — a Resolved row, and Open #10 narrowed rather than closed (the
`Ramadan` helpers still ignore the offset).
