+++ app/src/main/java/com/arshadshah/nimaz/core/di/RepositoryModule.kt
@@ -59,6 +59,7 @@ import com.arshadshah.nimaz.domain.repository.settings.LocationSettings
 import com.arshadshah.nimaz.domain.repository.settings.QuranPreferences
 import com.arshadshah.nimaz.domain.repository.settings.TasbihSettings
 import com.arshadshah.nimaz.domain.repository.settings.MoreSettings
+import com.arshadshah.nimaz.domain.repository.settings.HijriSettings
 import com.arshadshah.nimaz.domain.repository.settings.ZakatSettings
 import com.arshadshah.nimaz.domain.repository.TafseerRepository
 import com.arshadshah.nimaz.domain.repository.TasbihRepository
@@ -467,6 +468,9 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindZakatSettings(impl: PreferencesDataStore): ZakatSettings
+
+    @Binds
+    abstract fun bindHijriSettings(impl: PreferencesDataStore): HijriSettings
 
     @Binds
     @Singleton
