+++ app/src/test/java/com/arshadshah/nimaz/presentation/viewmodel/tracker/FastingViewModelTest.kt
@@ -1,6 +1,8 @@
 package com.arshadshah.nimaz.presentation.viewmodel.tracker
 
 import com.arshadshah.nimaz.core.time.FakeTodayProvider
+import com.arshadshah.nimaz.domain.usecase.fasting.GetDaysUntilAyyamAlBeedUseCase
+import com.arshadshah.nimaz.presentation.viewmodel.FakeHijriSettings
 import com.arshadshah.nimaz.presentation.viewmodel.FakeZakatSettings
 import com.arshadshah.nimaz.core.monitoring.RecordingTelemetry
 
@@ -84,6 +86,8 @@ class FastingViewModelTest {
             buildFastingUseCases(repository),
             buildPrayerUseCases(prayers),
             FakeTodayProvider(LocalDate.now()),
+            GetDaysUntilAyyamAlBeedUseCase(FakeTodayProvider(LocalDate.now())),
+            FakeHijriSettings(),
             FakeZakatSettings(),
             RecordingTelemetry(),
         )
